[refactor] team: GroupMenu 드롭다운 메뉴 컴포넌트 분리
