+++ src/app/(team)/team/[teamid]/page.tsx
@@ -1,7 +1,7 @@
 import { cookies } from 'next/headers';
 import { getGroupById } from '@/lib/apis/group';
 import { notFound } from 'next/navigation';
-import TeamBanner from '@/app/(team)/team/_components/TeamBanner';
+import TeamBanner from '@/app/(team)/team/_components/TeamBanner/TeamBanner';
 import TaskListBar from '@/app/(team)/team/_components/TaskListBar';
 
 export default async function TeamPage({
@@ -18,7 +18,7 @@ export default async function TeamPage({
   }
 
   return (
-    <div className="flex h-screen w-full flex-col items-center gap-6 p-6">
+    <div className="flex w-full flex-col items-center gap-6 p-6">
       <TeamBanner group={groupData} userId={Number(userId)} />
       <TaskListBar />
     </div>
