+++ src/app/(team)/team/_components/TeamBanner/TeamBanner.tsx
@@ -1,0 +1,42 @@
+'use client';
+import Image from 'next/image';
+import GroupMenu from '@/app/(team)/team/_components/TeamBanner/GroupMenu';
+import GradientScrollable from '@/components/common/Scroll/GradientScrollable';
+import { useIsAdmin } from '@/hooks/useIsAdmin';
+import { GroupResponse } from '@/lib/apis/group/type';
+import {
+  teamBannerWrapperStyle,
+  teamBannerImgStyle,
+  teamBannerTitleStyle,
+} from '@/app/(team)/team/_components/styles';
+
+const TeamBanner = ({
+  group,
+  userId,
+}: {
+  group: GroupResponse;
+  userId: number;
+}) => {
+  const isAdmin = useIsAdmin({ membersData: group.members, userId });
+
+  return (
+    <div className={`${teamBannerWrapperStyle}`}>
+      <Image
+        src="/image/team_banner_pattern.svg"
+        alt="팀 배너 배경 패턴 이미지"
+        width={181}
+        height={64}
+        priority
+        className={`${teamBannerImgStyle}`}
+      />
+
+      <div className={`${teamBannerTitleStyle}`}>
+        <GradientScrollable>{group.name}</GradientScrollable>
+      </div>
+
+      {isAdmin && <GroupMenu group={group} />}
+    </div>
+  );
+};
+
+export default TeamBanner;
